trying to make cordova work

+++ www/app/controllers/listController.js
@@ -2,7 +2,7 @@
  * Created by ejimenez on 5/26/2014.
  */
 
-app.controller('listController', ['$scope','$rootScope', 'volunteerService', '$location', function ($scope, $rootScope, service, $location) {
+app.controller('listController', ['$scope','$rootScope', 'volunteerService', '$location', 'cordovaService', function ($scope, $rootScope, service, $location, cordovaService) {
     $rootScope.$on("$routeChangeStart", function(){
         $rootScope.loading = true;
     });
@@ -17,7 +17,9 @@ app.controller('listController', ['$scope','$rootScope', 'volunteerService', '$l
 
     $scope.removeEvent = function (event, evt) {
         evt.preventDefault();
-        service.removeVolunteerEvents(event);
+        cordovaService.confirm("Do you want to remove event?", "Remove Event", function() {
+            service.removeVolunteerEvents(event);
+        })
     };
 
 //    $scope.groceries = [];
